move lenenc_int from util.py to connections.py
