+++ pymysql/connections.py
@@ -22,7 +22,7 @@ from .converters import (
     escape_item, encoders, decoders, escape_string, through)
 from .cursors import Cursor
 from .optionfile import Parser
-from .util import byte2int, int2byte, lenenc_int
+from .util import byte2int, int2byte
 from . import err
 
 try:
@@ -210,6 +210,20 @@ def _hash_password_323(password):
 def pack_int24(n):
     return struct.pack('<I', n)[:3]
 
+# https://dev.mysql.com/doc/internals/en/integer.html#packet-Protocol::LengthEncodedInteger
+def lenenc_int(i):
+    if (i < 0):
+        raise ValueError("Encoding %d is less than 0 - no representation in LengthEncodedInteger" % i)
+    elif (i < 0xfb):
+        return int2byte(i)
+    elif (i < (1 << 16)):
+        return b'\xfc' + struct.pack('<H', i)
+    elif (i < (1 << 24)):
+        return b'\xfd' + struct.pack('<I', i)[:3]
+    elif (i < (1 << 64)):
+        return b'\xfe' + struct.pack('<Q', i)
+    else:
+        raise ValueError("Encoding %x is larger than %x - no representation in LengthEncodedInteger" % (i, (1 << 64)))
 
 class MysqlPacket(object):
     """Representation of a MySQL response packet.
