+++ pymysql/util.py
@@ -18,17 +18,3 @@ def join_bytes(bs):
             rv += b
         return rv
 
-# https://dev.mysql.com/doc/internals/en/integer.html#packet-Protocol::LengthEncodedInteger
-def lenenc_int(i):
-    if (i < 0):
-        raise ValueError("Encoding %d is less than 0 - no representation in LengthEncodedInteger" % i)
-    elif (i < 0xfb):
-        return int2byte(i)
-    elif (i < (1 << 16)):
-        return b'\xfc' + struct.pack('<H', i)
-    elif (i < (1 << 24)):
-        return b'\xfd' + struct.pack('<I', i)[:3]
-    elif (i < (1 << 64)):
-        return b'\xfe' + struct.pack('<Q', i)
-    else:
-        raise ValueError("Encoding %x is larger than %x - no representation in LengthEncodedInteger" % (i, (1 << 64)))
